Fixed a few warnings.
Rev: bin/mkpeep.pike:1.18
Rev: src/peep.c:1.34
Rev: src/peep.h:1.6

+++ src/peep.h
@@ -1,5 +1,5 @@
 /*
- * $Id: peep.h,v 1.5 2000/04/20 02:41:45 hubbe Exp $
+ * $Id: peep.h,v 1.6 2000/08/14 17:18:06 grubba Exp $
  */
 #ifndef PEEP_H
 #define PEEP_H
@@ -11,22 +11,22 @@ extern dynamic_buffer instrbuf;
 struct p_instr_s;
 void init_bytecode(void);
 void exit_bytecode(void);
-int insert_opcode2(unsigned int f,
-		  INT32 b,
-		  INT32 c,
-		  INT32 current_line,
-		  struct pike_string *current_file);
-int insert_opcode1(unsigned int f,
-		  INT32 b,
-		  INT32 current_line,
-		  struct pike_string *current_file);
-int insert_opcode0(int f,int current_line, struct pike_string *current_file);
+ptrdiff_t insert_opcode2(unsigned int f,
+			 INT32 b,
+			 INT32 c,
+			 INT32 current_line,
+			 struct pike_string *current_file);
+ptrdiff_t insert_opcode1(unsigned int f,
+			 INT32 b,
+			 INT32 current_line,
+			 struct pike_string *current_file);
+ptrdiff_t insert_opcode0(int f,int current_line, struct pike_string *current_file);
 void update_arg(int instr,INT32 arg);
 void ins_f_byte(unsigned int b);
 void assemble(void);
-int insopt2(int f, INT32 a, INT32 b, int cl, struct pike_string *cf);
-int insopt1(int f, INT32 a, int cl, struct pike_string *cf);
-int insopt0(int f, int cl, struct pike_string *cf);
+ptrdiff_t insopt2(int f, INT32 a, INT32 b, int cl, struct pike_string *cf);
+ptrdiff_t insopt1(int f, INT32 a, int cl, struct pike_string *cf);
+ptrdiff_t insopt0(int f, int cl, struct pike_string *cf);
 /* Prototypes end here */
 
 #endif
